Added energy system and beacon
- Added energy beacon for evangelion units.
The evangelion units inside the supply radius can operate without energy restrictions, and recharge they battery at a rate of 3.5s/s.  Must be activated through ACS. Has 150 hp.

- Added energy requirements for evangelion units.
Similarly to synchronization; moving, jumping or using a weapon or skill requires and consume energy every second. The EVA01 starts with 5 minutes of energy.

- Added hud piece for energy meter.
When the eva unit is not in a charging area, a battery meter will show up the remaining operational time.

- Internal: Updated resting/moving system.
- Internal: Updated moving speed vector magnitude read function.